requirements.txt
openai
tiktoken
langchain
lancedb
flask
GitPython
flask_cors
asyncio
python-dotenv
pandas